+++ client/src/components/Settings/Dns/Config/Form.tsx
@@ -101,6 +101,10 @@ const Form = ({ processing, initialValues, onSubmit }: Props) => {
             label: t('null_ip'),
         },
         {
+            value: BLOCKING_MODES.noerror,
+            label: t('noerror'),
+        },
+        {
             value: BLOCKING_MODES.custom_ip,
             label: t('custom_ip'),
         },
@@ -111,6 +115,7 @@ const Form = ({ processing, initialValues, onSubmit }: Props) => {
         t(`blocking_mode_refused`),
         t(`blocking_mode_nxdomain`),
         t(`blocking_mode_null_ip`),
+        t(`blocking_mode_noerror`),
         t(`blocking_mode_custom_ip`),
     ];