+++ client_v2/src/components/DnsSettings/helpers.ts
@@ -33,6 +33,8 @@ export const getBlockingModeSummary = (mode: string): string => {
             return 'REFUSED';
         case BLOCKING_MODES.nxdomain:
             return 'NXDOMAIN';
+        case BLOCKING_MODES.noerror:
+            return 'NOERROR';
         case BLOCKING_MODES.null_ip:
             return intl.getMessage('dns_blocking_mode_null_ip');
         case BLOCKING_MODES.custom_ip:
@@ -84,6 +86,11 @@ export const getBlockingModeOptions = () => {
             text: intl.getMessage('dns_blocking_mode_null_ip'),
             value: BLOCKING_MODES.null_ip,
             description: intl.getMessage('dns_blocking_mode_null_ip_desc'),
+        },
+        {
+            text: intl.getMessage('dns_blocking_mode_noerror'),
+            value: BLOCKING_MODES.noerror,
+            description: intl.getMessage('dns_blocking_mode_noerror_desc'),
         },
         {
             text: intl.getMessage('dns_blocking_mode_custom_ip'),